Metric: Add time since last RAFT election and its reason
Export a new metric to show time since last
RAFT election for NB and SB DBs occured in seconds
and what was the reason for the election.

This metric will only set a valid value for the RAFT
leader and the followers will set a Not a Number (NaN).
Unfortunately, the way the current metrics are designed
to capture cluster/show metrics, we cannot easily
enable / disable this metric only when the OVSDB is
leader so we need to set followers value to NaN to
stop their consumption.

Consumers of this metric should only use this metric
when the server_role label is leader.

This metric can be used to inform the cluster admin of
the reason for the last election and how long ago it
occured.

Cardinality is acceptable.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>