Metrics: add config duration monitor
Record duration between a time submitted to func
RecordConfigDuration and when all nodes have this
configuration applied. Exposing a histogram to
expose this information to a user.

Disabled by default and if enabled, it will automatically
disable itself if node count is greater than 100 to protect
SB DB. This can be increased in the future following scale
testing.

The option to record is available only within periods of
time to limit the load on OVN and when not available its
a no-op.
The period is calculated according to the number of nodes
within the cluster - a second for each node.

Also, exposed a histograms to detect hung ovn-controllers.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>