Add utility DeleteLogicalRouterStaticRoutesWithPredicateOps
This commit adds DeleteLogicalRouterStaticRoutesWithPredicateOps
to libovsdbops package. Note that we already had
DeleteLogicalRouterStaticRoutesWithPredicate and now we want to
just have a util that returns the ops so that we can batch this
when doing a deletion for EIP.

NOTE: This commit only introduces the utility, its use comes down
in future commit.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>