Refactor how node IP manager gets its IP addresses
The reason for this change is, before we just accepted all
IPs seen on a node that met a couple of conditions without
ever checking the status of the link the address is associated
with. For example, if a link is done.

This change changes the behavour of node ip. Previously,
node IP would push v4 and v6 addresses to host-address
annotation, but now we only push addresses if the
IP family is supported.

Also, before, we just ensured we excluded link local addresses
but now we ensure the addresses have scope global.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>