Depreciate IPv6 host masq IP
The host masq IP is used as src IP when a workload
residing in the host net ns attempts to connect to
a svc IP.

Users may not config any routes that match a particular
destination.

Then linux will use the follow src addr selection algo
for IPv6:

1. Prefer same address.
2. Prefer appropriate scope.
3. Avoid deprecated addresses.
4. Prefer home addresses.
5. Prefer outgoing interface
6. Prefer matching label.
7. Prefer public addresses.
8. Use longest matching prefix.

If theres a tie break, linux uses the latest address added,
and that can end up being the host masq IP.

We want to only use the host masq IP if the dst is a
a service IP and never use it otherwise.

We cannot change the scope of the address because the host
IP masq IP falls within the global scope range as defined
by IANA special registry and it wouldnt totally solve
the problem anyway, just improve it.

Depreciating the address was not meant for this purpose
but it achieve this goal as long as we have the correct
routes configured.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>