EIP E2Es: add multi NIC test cases for enable/disable egress
This PR adds two test cases that are related.
They are both attempting to trigger egress IP
reassignment by either disabling a node for egress or
unselecting a pod.

The first test case attempts to do this for non-OVN
managed networks [1] and the second case attempts to
do this for both OVN [2] and non-OVN managed networks.

[1] non-OVN managed networks means that OVN is not directly
controller an interface attached to whatever network.
[2] OVN managed means OVN is directly controlling this interface
attached to whatever network.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>