Link manager: collect errors and return when syncing
syncLink fn got an error return a year ago and
I think its correct to collect the errors while
processing and return them instead of logging.
This allows callers to handle the errors also.

Also, we would emit a log of success even if
adding an address failed. Stop that.

Continue only logging GARPs.

Renamed addressWanted to wantedAddress because
this follows the common adjective-noun pattern.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>